Throw API error when Request Body fails to parse
There is a bug in the implemenetation of
OpenAPIOperationController.apply() that means that it doesn't enforce a
required request body in the case where a request body is supplied but
is in an unsupported MIME type.  This adds the necessary else branch to
the logic to throw an error in the case of a present but unsupported
format request body.